1. 공통 코드 정리 2. User 상세 조회 기능 추가 3. Auth 메일 인증 수단 기능 추가

+++ api/todo/todo.ctrl.js
@@ -2,22 +2,28 @@ const Todo = require('../../models').Todo;
 
 const index = (req, res) => {
   req.query.limit = req.query.limit || 10;
+
+  const creator_id = req.user.id;
   const limit = Number(req.query.limit);
   if (Number.isNaN(limit)) return res.status(400).end();
 
   Todo
-    .findAll({ limit: limit })
+    .findAll({
+      where: { creator_id },
+      limit: limit
+    })
     .then(todos => {
       res.json(todos);
     });
 };
 
 const show = (req, res) => {
+  const creator_id = req.user.id;
   const id = Number(req.params.id);
   if (Number.isNaN(id)) return res.status(400).end();
 
   Todo
-    .findOne({ where: { id }})
+    .findOne({ where: { creator_id, id }})
     .then(todo => {
       if (!todo) return res.status(404).end();
       res.json(todo);
@@ -25,24 +31,26 @@ const show = (req, res) => {
 };
 
 const destroy = (req, res) => {
+  const creator_id = req.user.id;
   const id = Number(req.params.id);
   if (Number.isNaN(id)) return res.status(400).end();
 
   Todo
-    .destroy({ where: { id } })
+    .destroy({ where: { creator_id, id } })
     .then(() => {
       res.status(204).end();
     });
 };
 
 const create = (req, res) => {
+  const creator_id = req.user.id;
   const title = req.body.title;
   if (!title) return res.status(400).end();
 
   const message = req.body.message || '';
 
   Todo
-    .create({ title, message })
+    .create({ title, message, creator_id })
     .then(todo => {
       res.status(201).json(todo);
     })
